Enable optionally not depending on the std library.
Involved refactoring the equivalence-classes data structures to be generic and
public, so they do not depend on std.  Now these generic aspects are available
even without any features enabled and no-std users can impl their own types for
them.

Required adding a generic type parameter to the Recur trait, so that it can be
impl'ed outside of the crate for the Equiv type of the crate with an outside
custom type as type argument, as needed by no-std or customizing users.

Required some module/item namespace and visibility restructuring.  This
triggered some lints and these were resolved, including doc comments, inline,
and non_exhaustive.