Support choosing different (P)RNG crates, including no_std ones.
fastrand can only be used with the "std" feature.

Add oorandom and wyhash support, both of which are no_std.

nanorand, which can be no_std, was evaluated but not added because its variety
of features is more complicated to expose, and fastrand already offers a WyRand
using thread-local storage that requires std, and wyhash already offers a WyRand
that does not use TLS and is no_std, and oorandom already offers a PCG that is
no_std.

The use of an RNG is now generalized as a simple trait (but is not parameterized
as generics), and the chosen implementation is determined by package feature and
is provided as a module.  It is easy to add support for other (P)RNG crates.