Refactor generic parameters to be grouped into structured traits.
This improves the API by better organizing how generic arguments are given and
by eliminating the need for artificial generic parameters and awkward
constraints.

The constants for limits and capacities are now also associated with these
traits and can now be redefined by users.

The (P)RNG type is also now generic and multiple can now be used (unusual), and
one is default by priority.