Add benchmarks and an overview of their results.
The default value of vecstack::Params::INITIAL_CAPACITY is reduced to what seems
a little more reasonable and what allows the benchmarks to not take too long
with the other values they derive from this.

PairChainMaker::degenerate_cyclic is tweaked to cause the same amount of
recursions as degenerate_dag for the same depth.

`#[must_use]` is added to a couple functions, to resolve lints when using the
"nightly" compiler (as the benches require).