Enhance to support multi-way comparisons and generic error types.
Now graphs can be given an ordering, in addition to equivalence, when desired.

Both methods of DescendMode now can return an error to cause aborting or can
return false to cause skipping.  The RecurStack::recur method now can return an
error to cause aborting.

The new RecurStack::IndexIter associated type enables controlling the order that
edge indices are iterated, to enable different recursion-stack types to achieve
a consistent order that edges are recurred on, so that multi-way comparison
results are consistent with different recursion-stack types.

The impact of all these changes on performance is insignificant or maybe a
little better for VecStack.  The tail-call-elimination handling of VecStack is
now reversed, so that comparison results are consistent with CallStack, which
swaps the benchmark results for lists versus inverted lists.  A few additional
elaborations are added to the interpretations of the benchmarks.