Handle very-wide shapes with VecStack with much less memory.
Previously, all edges of a node were first pushed on to a VecStack before
recurring on them, which consumed excessive memory for very-wide shapes.  Now, a
new EdgesIter is what is pushed on, so that edges are gotten lazily, which
limits the consumption to a small constant per depth level.

The recently-added IndexIter is now removed and subsumed by EdgesIter.

The logic of ticker decrementing and limiting is changed slightly.  This enables
separating this from the logic of iterating edges.

A new wide test node type, and integration test using it, are added, because it
exercises very-wide degenerate shapes that require this commit in order to be
handled efficiently.

A new lazy test node type, and integration test using it, are added, because it
is useful to exercise the algorithm with logically-big shapes that do not
consume memory, which is useful to more easily gauge how much memory other
aspects consume.

The performance for some benchmarks has decreased somewhat, but there might be
ways to improve it later.  Being robust with very-wide shapes seems more
important.