Change to take input nodes by value, to avoid unnecessary clones.
While it now requires an initial extra clone, for some types, that is only
one-time, and the clones that were done for each recursion (that were introduced
for EdgesIter) are now eliminated.

The performance of most benchmarks has improved and regained the same speed as
before the use of EdgesIter.  The exception is VecStack is still not as fast as
before, but is now somewhat faster with this commit at least.