Simplify and improve ticker handling for Interleave.
Previously, a ticker was always decremented for leaf nodes, which complicated
the logic of choosing which phase should be used next.  Now a ticker is only
decremented when choosing a phase, which effectively only decrements it for
branch nodes now.

While the benchmarks show that the list shape is now processed somewhat slower,
when the length is short and when a precheck is not done, this is only because
the short list shape happened to get lucky previously regarding the pattern of
interleave phase alternations.  Now, the list shape is the same speed as the
inverted_list shape.  This is not a concern because, for the small shapes where
the slow-down is significant, the precheck should normally be done instead and
it will complete and so the slow-down is avoided, and for larger shapes the
slow-down is less significant.  This is an improvement because the interleave
phase is used when shared structure is expected, and now the detection of shared
structure is more consistent and is as frequent for all shapes as the chosen
parameters are supposed to cause.

For the wide shape, the performance is improved a little, due to no longer
decrementing a ticker for each of very many leaf nodes.