Rename to better reflect nature.
Rename the crate because it now covers more than just cycle- and depth-safety.

Rename the module that provides VecStack to wide_safe, because it is more
appropriate for wide shapes than for deep shapes.

Rename the type that provides recursion to RecurMode, because data structures
other than stacks may be used.