Improve handling of unstable features and of when they stabilize.
Add a build script that automatically detects select compiler/language/library
features of Rust and sets new configuration options for use by the new
conditional-compilation handling in the source code.  The functionality in this
build script is currently only a prototype of what should probably be moved into
its own crate.  The tests_utils crate needs this feature detection as well, and,
as a temporary convenience for now, uses this same build-script file by simply
symlink'ing to it, but this should be made into its own separate file.

Now the source code is able to cause indicative warnings if/when a feature
becomes stabilized in future versions of Rust, while still building
successfully, so we can see that our previous workarounds can now be deprecated
and then removed.

There are some other minor changes that enable building with the unstable
features activated and used, instead of using our workarounds.  But this ability
is currently not exposed, but it probably should be in the future.

Another benefit of the new automatically-set configuration options is that the
benches are now automatically conditionally deactivated with the stable or beta
versions of rustc (i.e. versions that are not nightly nor dev), which is
especially helpful with e.g. IDE tools that like to give --all-targets.  Now,
the stable and beta versions can have --benches given, or even do `cargo bench`,
and still build and run successfully.