Add "anticipate" package-feature that uses different Rust-features.
When activated, unstable Rust-features are used like they would be if adopted,
instead of the default workarounds.

Named "anticipate" instead of "unstable", because the used Rust features could
become stable in the future and would still be used as such by this
package-feature, and using them either way will always fit the concept of
anticipation of future breaking changes.