Eliminate the only remaining hint of user-caused panics.
Remove Node::amount_edges and change Node::get_edge to return Option, so that
users' implementations do not involve unreachable panic cases.  While the crate
previously would never have caused these cases, it feels cleaner to help users
not have them by construction and this helps them avoid causing panics
themselves if, in some very-unusual situation, they were to directly use the
crate's trait methods.

This also simplifies users' impl of the Node trait.  The internal RangeIter is
eliminated which is another simplification.

The benchmarks' performance is the same except the wide_safe variant is somewhat
slower with great-depth shapes but this does not seem like a concern because it
is intended more for great-width shapes (which usually do not have great
depth) (whereas there is the deep_safe variant that is performant for great
depth).